refactor: feature폴더에서 library>_component 폴더로 이동
