+++ src/app/library/detail/[id]/_components/ContentComponent.tsx
