+++ src/app/library/detail/[id]/_components/PaginationControl.tsx
