+++ src/app/library/detail/[id]/page.tsx
@@ -4,8 +4,8 @@ import { useGetContent } from '@/hooks/stories/useGetContent';
 import { useGetStory } from '@/hooks/stories/useGetStory';
 import { useParams } from 'next/navigation';
 import { useEffect, useState } from 'react';
-import ContentComponent from '@/components/feature/library/ContentComponent';
-import { PaginationControl } from '@/components/feature/library/PaginationControl';
+import ContentComponent from '@/app/library/detail/[id]/_components/ContentComponent';
+import { PaginationControl } from '@/app/library/detail/[id]/_components/PaginationControl';
 import { TeamUserRole } from '@/types/teamUserRole';
 import WritableUserModal from './_components/WritableUserModal';
 import { StoryWriteOrApproveModalProviders } from '@/providers/StoryWriteOrApproveModalProviders';
