--- src/components/common/Select.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Select = () => {
-  return <div>Select</div>;
-};
-
-export default Select;
